Backend/production/requirements.txt
# Production Requirements for RL-Based Text Optimization Framework
# Last updated: September 30, 2025

# Core ML and RL libraries
torch>=1.13.0
gymnasium>=0.26.0
stable-baselines3[extra]>=1.7.0
numpy>=1.21.0,<2.0.0

# NLP and Language Processing
sentence-transformers>=2.2.0
transformers>=4.20.0
spacy>=3.4.0
nltk>=3.8.0
tokenizers>=0.13.0
huggingface-hub>=0.15.0

# Data Science and Analysis
pandas>=1.5.0
scikit-learn>=1.1.0
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.11.0

# API and Web Framework
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
jinja2>=3.1.0
markupsafe>=2.1.0
itsdangerous>=2.1.0
click>=8.1.0
blinker>=1.6.0

# HTTP and Networking
requests>=2.28.0
urllib3>=1.26.0
certifi>=2022.12.0

# Gemini API Support
google-generativeai>=0.3.0

# System and Performance
psutil>=5.9.0
tqdm>=4.64.0
joblib>=1.1.0
threadpoolctl>=3.1.0

# JSON and Data Handling
jsonschema>=4.17.0
orjson>=3.8.0  # Fast JSON processing

# Development and Testing (optional but recommended)
pytest>=7.1.0
pytest-asyncio>=0.21.0
black>=22.10.0
flake8>=5.0.0

# GPU Acceleration Support
# These are automatically handled by PyTorch installation:
# - CUDA toolkit (install separately if needed)
# - cuDNN (install separately if needed)
# For CPU-only deployment, PyTorch CPU version will be used automatically

# System Dependencies (install via system package manager):
# Ubuntu/Debian: sudo apt-get install python3-dev build-essential
# CentOS/RHEL: sudo yum install python3-devel gcc gcc-c++
# macOS: xcode-select --install
# Windows: Microsoft Visual C++ Build Tools